v0.4.2

+++ Maid/Maid.py
@@ -1,41 +1,66 @@
 from discord import File,Member
 from discord.ext import commands
 from discord.ext.commands import MemberConverter
-import Okari
+import Okari, ExpSys
 import Memes_name_subject_to_change as mem
 import os
 
 bot = commands.Bot(command_prefix='!', description='Сomitete System')
 
+async def is_owner(ctx):
+    return ctx.author.id == 269860812355665921
+
+def is_testserver(ctx):
+    return ctx.guild.id == 198889006006665216
+
+infochan=198892670456692737
 
 @bot.event
 async def on_ready():
+    print('ExpSys --------- '+ExpSys.init())
     print('Logged in as')
     print(bot.user.name)
-    print(bot.user.id)
-    print('------')
+    print(bot.user.id)   
 
 #BotCommands
+@bot.command(name="@integrate")
+@commands.check(is_owner)
+async def integrate(ctx,*args):
+    for i in ctx.guild.members:
+        path=Okari.AddMember(i)
+        os.remove(path)  
+        ExpSys.AddMem(i.id)
+    await ctx.send("Done")
+
 @bot.command(name="@addmem")
+@commands.check(is_owner)
 async def addmem(ctx,*args):
     if args:
         author=await MemberConverter().convert(ctx,args[0])
     else:
         author=ctx.author
+    ExpSys.AddMem(ctx.author.id)
     path=Okari.AddMember(author)
     file=File(path,filename="Member.png")
     await ctx.send(file=file)
-    os.remove(path)    
+    os.remove(path)  
+
+@bot.command(name="@addexp")
+@commands.check(is_owner)
+async def addexp(ctx,arg1,arg2):
+    author=await MemberConverter().convert(ctx,arg1)
+    await ExpSys.AddExp(author.id,arg2,ctx.channel)
 
 @bot.command(name="@lostmem")
+@commands.check(is_owner)
 async def lostmem(ctx):
+    ExpSys.DelMem(ctx.author.id)
     path=Okari.LostMember(ctx.author)
     file=File(path,filename="LostMem.png")
     await ctx.send(file=file)
     os.remove(path)
 
 #memes
-
 @bot.command(name="memes")
 async def memes(ctx,*args):
     if os.path.exists('Maid/src/Images/memes/{}.json'.format(args[0])):
@@ -49,23 +74,26 @@ async def memes(ctx,*args):
 #MemberEvents
 @bot.event
 async def on_member_join(mem):
-    path=Okari.AddMember(mem)
-    file=File(path,filename="MemJoin.png")
-    await mem.guild.get_channel(566000934493224962).send(file=file)
-    os.remove(path)
+    if is_testserver(mem):
+        path=Okari.AddMember(mem)
+        ExpSys.AddMem(mem.id)
+        file=File(path,filename="MemJoin.png")
+        await mem.guild.get_channel(infochan).send(file=file)
+        os.remove(path)
 
 @bot.event
 async def on_member_remove(mem):
-    path=Okari.LostMember(mem)
-    file=File(path,filename="MemRemove.png")
-    await mem.guild.get_channel(566000934493224962).send(file=file)
-    os.remove(path)
+    if is_testserver(mem):
+        path=Okari.LostMember(mem)
+        ExpSys.DelMem(mem.id)
+        file=File(path,filename="MemRemove.png")
+        await mem.guild.get_channel(infochan).send(file=file)
+        os.remove(path)
 
-'''
 @bot.event
 async def on_message(message):
-    if message.author.id != bot.user.id:
-        file=discord.File(AddMember(message.author),filename="Newmember.png")
-        await message.channel.send(file=file)
-'''
+    await bot.process_commands(message)
+    if is_testserver(message):
+        await ExpSys.AddExp(message.author.id,len(message.content)/10,message.channel)
+        
 bot.run(open('Maid/AccessToken','r').read())
